Im CMAKE gibt es nun die Option MATRIX_DOUBLE_PRECISION. Wenn auf "ON", dann werden die Matrizen in double gespeichert. Üblich ist float.

+++ src/LinAlg/sparsestructure.cc
@@ -49,7 +49,7 @@ ostream& operator<<(ostream &s, const SparseStructure& A)
 
 void SparseStructure::statistics(ostream &s) const
 {
-  s << n() << " " << ntotal() << "  " <<  ntotal()/static_cast<float>(n());
+  s << n() << " " << ntotal() << "  " <<  ntotal()/static_cast<double>(n());
 }
 
 /*----------------------------------------------*/
